requirements.txt
requests
schedule
python-nmap
pydantic
pydantic-settings
aiofiles
google-cloud-bigquery
PyYAML>=6.0
